+++ src/components/icons/instagram/index.tsx
