+++ src/components/icons/linkedin/index.tsx
