+++ src/pages/_app.tsx
@@ -1,4 +1,4 @@
-import "@src/styles/globals.css";
+import "@src/styles/globals.scss";
 
 import App from "../screens/app";
